moved test classes into their own package
